refactor(engine): moved svg color replace, removed image preloading, it's no more necessary
